Rename RSError -> RSNetworkError
