+++ frontend-react/src/utils/RSNetworkError.ts
